+++ components/buildercontrols/AppBuilderVerticalTabContent.tsx
